requirements/base.txt
django
psycopg2
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
django-ckeditor==6.1.0
djangorestframework-simplejwt
